requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
motor==3.2.0
pydantic==2.0.2
pydantic_core==2.1.2
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
xmltodict==0.13.0
